:rotating_light: :bug: Fix JB2008 test in 32-bin envs. again

+++ src/earth/space_indices/space_indices.jl
@@ -227,7 +227,7 @@ function _parse_dtcfile(path::AbstractString)
             end
 
             # Compute the Julian day at midnight of the day.
-            year = parse(Int64, tokens[2])
+            year = parse(Int,     tokens[2])
             doy  = parse(Float64, tokens[3])
 
             JD_0h = DatetoJD(year, 1, 1, 0, 0, 0) - 1 + doy
